Split up column tests.
