+++ db/tests/columns/test_base.py
@@ -1,0 +1,245 @@
+import re
+
+import pytest
+from sqlalchemy import String, Integer, ForeignKey, Table, MetaData, Numeric
+
+from db.columns.base import MathesarColumn
+from db.columns.defaults import DEFAULT_COLUMNS
+from db.columns.utils import get_default_mathesar_column_list
+from db.tests.types import fixtures
+from db.types import email
+
+
+engine_with_types = fixtures.engine_with_types
+temporary_testing_schema = fixtures.temporary_testing_schema
+engine_email_type = fixtures.engine_email_type
+
+
+def init_column(*args, **kwargs):
+    return MathesarColumn(*args, **kwargs)
+
+
+def from_column_column(*args, **kwargs):
+    """
+    This creates a condition to check that the MathesarColumn.from_column
+    class method returns the same value of the original __init__ for a given
+    input.
+    """
+    col = MathesarColumn(*args, **kwargs)
+    return MathesarColumn.from_column(col)
+
+
+column_builder_list = [init_column, from_column_column]
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_name(column_builder):
+    name = "mycol"
+    col = column_builder(name, String)
+    assert col.name == name
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_sa_type(column_builder):
+    sa_type = String
+    col = column_builder("anycol", sa_type)
+    actual_cls = col.type.__class__
+    assert actual_cls == sa_type
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_default_not_primary_key(column_builder):
+    col = column_builder("anycol", String)
+    assert not col.primary_key
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_primary_key_true(column_builder):
+    col = column_builder("anycol", String, primary_key=True)
+    assert col.primary_key
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_default_nullable(column_builder):
+    col = column_builder("a_col", String)
+    assert col.nullable
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_nullable_false(column_builder):
+    col = column_builder("a_col", String, nullable=False)
+    assert not col.nullable
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_foreign_keys_empty(column_builder):
+    col = column_builder("some_col", String)
+    assert not col.foreign_keys
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_non_empty_foreign_keys(column_builder):
+    fk_target = "some_schema.some_table.a_column"
+    col = column_builder(
+        "anew_col", String, foreign_keys={ForeignKey(fk_target)},
+    )
+    fk_names = [fk.target_fullname for fk in col.foreign_keys]
+    assert len(fk_names) == 1 and fk_names[0] == fk_target
+
+
+@pytest.mark.parametrize("column_builder", column_builder_list)
+def test_MC_inits_with_server_default_none(column_builder):
+    col = column_builder("some_col", String)
+    assert col.server_default is None
+
+
+def test_MC_is_default_when_true():
+    for default_col in get_default_mathesar_column_list():
+        assert default_col.is_default
+
+
+def test_MC_is_default_when_false_for_name():
+    for default_col in DEFAULT_COLUMNS:
+        dc_definition = DEFAULT_COLUMNS[default_col]
+        col = MathesarColumn(
+            "definitely_not_a_default",
+            dc_definition["sa_type"],
+            primary_key=dc_definition.get("primary_key", False),
+            nullable=dc_definition.get("nullable", True),
+        )
+        assert not col.is_default
+
+
+def test_MC_is_default_when_false_for_type():
+    for default_col in DEFAULT_COLUMNS:
+        dc_definition = DEFAULT_COLUMNS[default_col]
+        changed_type = Integer if dc_definition["sa_type"] == String else String
+        col = MathesarColumn(
+            default_col,
+            changed_type,
+            primary_key=dc_definition.get("primary_key", False),
+            nullable=dc_definition.get("nullable", True),
+        )
+        assert not col.is_default
+
+
+def test_MC_is_default_when_false_for_pk():
+    for default_col in DEFAULT_COLUMNS:
+        dc_definition = DEFAULT_COLUMNS[default_col]
+        not_pk = not dc_definition.get("primary_key", False),
+        col = MathesarColumn(
+            default_col,
+            dc_definition["sa_type"],
+            primary_key=not_pk,
+            nullable=dc_definition.get("nullable", True),
+        )
+        assert not col.is_default
+
+
+def test_MC_valid_target_types_no_engine():
+    mc = MathesarColumn('testable_col', String)
+    assert mc.valid_target_types is None
+
+
+def test_MC_valid_target_types_default_engine(engine):
+    mc = MathesarColumn('testable_col', String)
+    mc.add_engine(engine)
+    assert "VARCHAR" in mc.valid_target_types
+
+
+def test_MC_valid_target_types_custom_engine(engine_with_types):
+    mc = MathesarColumn('testable_col', String)
+    mc.add_engine(engine_with_types)
+    assert "MATHESAR_TYPES.EMAIL" in mc.valid_target_types
+
+
+def test_MC_column_index_when_no_engine():
+    mc = MathesarColumn('testable_col', String)
+    assert mc.column_index is None
+
+
+def test_MC_column_index_when_no_table(engine):
+    mc = MathesarColumn('testable_col', String)
+    mc.add_engine(engine)
+    assert mc.column_index is None
+
+
+def test_MC_column_index_when_no_db_table(engine):
+    mc = MathesarColumn('testable_col', String)
+    mc.add_engine(engine)
+    table = Table('atable', MetaData(), mc)
+    assert mc.table == table and mc.column_index is None
+
+
+def test_MC_column_index_single(engine_with_schema):
+    engine, schema = engine_with_schema
+    mc = MathesarColumn('testable_col', String)
+    mc.add_engine(engine)
+    metadata = MetaData(bind=engine, schema=schema)
+    Table('asupertable', metadata, mc).create()
+    assert mc.column_index == 0
+
+
+def test_MC_column_index_multiple(engine_with_schema):
+    engine, schema = engine_with_schema
+    mc_1 = MathesarColumn('testable_col', String)
+    mc_2 = MathesarColumn('testable_col2', String)
+    mc_1.add_engine(engine)
+    mc_2.add_engine(engine)
+    metadata = MetaData(bind=engine, schema=schema)
+    Table('asupertable', metadata, mc_1, mc_2).create()
+    assert mc_1.column_index == 0
+    assert mc_2.column_index == 1
+
+
+def test_MC_plain_type_no_opts(engine):
+    mc = MathesarColumn('acolumn', String)
+    mc.add_engine(engine)
+    assert mc.plain_type == "VARCHAR"
+
+
+def test_MC_plain_type_no_opts_custom_type(engine_with_types):
+    mc = MathesarColumn('testable_col', email.Email)
+    mc.add_engine(engine_with_types)
+    assert mc.plain_type == "MATHESAR_TYPES.EMAIL"
+
+
+def test_MC_plain_type_numeric_opts(engine):
+    mc = MathesarColumn('testable_col', Numeric(5, 2))
+    mc.add_engine(engine)
+    assert mc.plain_type == "NUMERIC"
+
+
+def test_MC_type_options_no_opts(engine):
+    mc = MathesarColumn('testable_col', Numeric)
+    mc.add_engine(engine)
+    assert mc.type_options is None
+
+
+def test_MC_type_options(engine):
+    mc = MathesarColumn('testable_col', Numeric(5, 2))
+    mc.add_engine(engine)
+    assert mc.type_options == {'precision': 5, 'scale': 2}
+
+
+def get_mathesar_column_init_args():
+    init_code = MathesarColumn.__init__.__code__
+    return init_code.co_varnames[1:init_code.co_argcount]
+
+
+@pytest.mark.parametrize("mathesar_col_arg", get_mathesar_column_init_args())
+def test_test_columns_covers_MathesarColumn(mathesar_col_arg):
+    """
+    This is a meta-test to require at least one test for each __init__
+    arg of the column object. It is stupid, and only checks function
+    names. To get it to pass, make a test function with "test_MC_inits"
+    and <param> in the name. It is the responsibility of the implementer
+    of new arguments to MathesarColumn to ensure the tests they write are
+    valid.
+    """
+    test_funcs = [var_name for var_name in globals() if var_name[:4] == "test"]
+    pattern = re.compile(r"test_MC_inits\S*{}\S*".format(mathesar_col_arg))
+    number_tests = len(
+        [func for func in test_funcs if pattern.match(func) is not None]
+    )
+    assert number_tests > 0
